Fix issue where errors aren't reported for a new overlay
A new overlay was not being marked as an explicitly added
source, even if it's within the directory of an analysis
context. As a result, errors were not being reported.

I believe this only happens if there's no underlying file.
Therefore it seems difficult to reproduce in an IDE, but
it comes up in a stress test.

BUG=
R=brianwilkerson@google.com

Review URL: https://codereview.chromium.org/1753083003 .