Add zlib dependency for dart_bootstrap build
The Dartium build does not include io components in the libraries, which
are what pull in zlib for the non-dartium builds of the SDK.  But it does
build dart_bootstrap, which needs zlib.  It actually should not need to
build dart_bootstrap, since it is useless without io capabilities.

BUG=
R=kasperl@google.com

Review URL: https://codereview.chromium.org/1761233003 .