Fix path handling when creating summaries.
The path passed to "--packageMode" must be converted to an absolute path
because it is used to normalize file paths (which are automatically
converted to absolute paths by the PhysicalResourceProvider.

We should not call `resourceProvider.pathContext.absolute()`.  This
seems (by its name) like it would convert its argument to an absolute
path.  But in fact it simply prepends the path with
`resourceProvider.pathContext.current`, which is `.`, and this causes
Source objects to be created referring to non-canonical paths.

R=scheglov@google.com

Review URL: https://codereview.chromium.org/1762363002 .