Timing hooks to profile linters (#24548).
Adds a `TimedAstVisitor` that wraps and times base visitors.  The perf hit for collecting timing was neglible so I opted to NOT surface a way to not collect the data.  We can add that if needed.

Once this is published in a new analyzer, I can bake client support into (for example) the linter.

More context and a client mock here: https://github.com/dart-lang/sdk/issues/24548

BUG=
R=brianwilkerson@google.com

Review URL: https://codereview.chromium.org/1807723002 .