Use exact instruction names in the SSA tracer.
Before the mapping from HInstruction names to printed representations
used by the tracer was arbitrary.  Change it to use the HInstruction
name with a colon separating the arguments from the instruction name.
The first colon on the line is recognized by the tool as delimiting
the opcode.

Seeing the precise structure of the IR is more important than
having human-writable surface syntax.

BUG=
R=sra@google.com

Review URL: https://codereview.chromium.org/1819833002 .