Remove sanity check.
Apparently, we are not completely sane. At least, there is a regression
test that expects us to be insane. Presumably that test reflects some
real code in the wild that also causes analyzer to hit an insane path.

Since we do handle the case anyway, remove the asserts and add a TODO.

R=jmesserly@google.com

Review URL: https://codereview.chromium.org/1828043002 .