Fix static member access errors when using summaries.
When summaries are in use, the ClassElement instances seen by the resolver
are sometimes ClassElementHandle, not ClassElementImpl.  So it's not
safe for the resolver to do "is" checks against ClassElementImpl.  This
was causing the resolver to incorrectly report the error
StaticTypeWarningCode.INSTANCE_ACCESS_TO_STATIC_MEMBER when code
attempted to access a static member of a class that was resynthesized
from a summary.

R=scheglov@google.com

Review URL: https://codereview.chromium.org/1826853005 .