Use AbstractSingleUnitTest, so MockSdk in resynthesize tests.
When we prelink AST based summaries we need to be able to read
referenced sources. So, mock SDK element model is not enough. Plus it
was quite a lot of work in the past to fix this mock element model to
provide all the data we expect from it during summarizing.

One of the negative effects is increasing analysis time from 26 to 34
seconds. We can try to implement sharing SDK between tests later.

R=paulberry@google.com
BUG=

Review URL: https://codereview.chromium.org/1827293002 .