Issue 26096. Explicitly compute the result that ensures that directives are resolved in units.
In the past computing LIBRARY_ELEMENT2 also set elements for directives
in the defining unit AST. But when this unit is flushed and reparsed
we didn't set these directive elements again because LIBRARY_ELEMENT2
is valid. Now resolving directive ASTs to elements is an explicit
dependency.

R=brianwilkerson@google.com, paulberry@google.com
BUG= https://github.com/dart-lang/sdk/issues/26096

Review URL: https://codereview.chromium.org/1854863002 .