Fix service isolate shutdown for precompilation.
Release and debug precompilation VMs start the service isolate, but do
not wait for it to startup (it is there to support Observatory and not
for loading). The main isolate can finish and the VM begin to shutdown
before the service isolate runs main. Handle this case.

BUG=http://dartbug.com/26144
R=johnmccutchan@google.com

Review URL: https://codereview.chromium.org/1855133003 .