Skip synthetic initializer functions in AST-based type inference tests.
When checking that AST-based type inference works properly, skip
checking of the synthetic functions that we create to represent
variable initializer expressions.  These synthetic elements are not
part of the public API of the code being analyzed, so it's not
necessary for them to work properly right now (our current goal is to
restrict AST-based type inference to only types that are visible
through the public API).

R=scheglov@google.com

Review URL: https://codereview.chromium.org/1862203002 .