Dart VM: Fix an order-of-evaluation bug in async code.
Reading and writing local and static variables is a side effect that
might be observable from a Future, so such reads and writes cannot be
scheduled after the Future completes when they should be performed
before.

The same is true for throwing an exception.

Closes https://github.com/dart-lang/sdk/issues/26175

R=hausner@google.com
BUG=

Review URL: https://codereview.chromium.org/1861963007 .