Extract expression type computer into a separate class.
The 'evaluate' method grew into an unmanageable mess :-(

If this happens to be a performance bottleneck we could push all
extracted methods as local functionsm and fields as local variables.
But hopefully VM is smart enough to notice that the class has no
subclasses, so there is no difference. I don't see any change in
time of runnings resynthesize_ast_test.dart - it is less than
variation.

R=paulberry@google.com
BUG=

Review URL: https://codereview.chromium.org/1872103002 .