Fix AST type inference in the presence of `void`.
This required adding code to link.dart to properly resynthesize the
type `void` when it is found in another package's summary.  I also
added code to resynthesize `dynamic` and bottom, but I didn't add
tests for them, because:

(a) the code for `dynamic` is already tested by other tests (those
tests previously passed by coincidence, because `dynamic` was treated
as an undefined reference, which has type dynamic).

(b) the code for bottom is unreachable, since all of the inferred
types that we currently store implicitly convert bottom to `dynamic`.

R=brianwilkerson@google.com, scheglov@google.com

Review URL: https://codereview.chromium.org/1878663002 .