Rename libchrome_zip to libzip_dart in build.
A condition was added to the libzip target in Chromium, renaming
the output from libzip to libchrome_zip to avoid collisions
with builtin libzip.  We copied this to our libzip_dart target,
so there was a collision in the Dartium build with two products
called libchrome_zip.  Removing this lets our copy be called
libzip_dart.

BUG=
R=iposva@google.com, terry@google.com

Review URL: https://codereview.chromium.org/1883593002 .