Turn off summaries in context when building summary from AST.
When building summaries from ASTs, we only use the context to read
source files; we don't need it to provide an element model.  Therefore,
we can save some time by turning off the context's use of summaries,
because this prevents it from trying to eagerly deserialize the
summaries of dart:async and dart:core.

Saves a significant amount of time when building a summary of a small
package using ASTs.  For example, saves ~37% of time when building a
strong mode summary of the "yaml" package.

R=brianwilkerson@google.com

Review URL: https://codereview.chromium.org/1882263002 .