Clean up handling of declared/inferred types in ExecutableElementForLink
Reworked the `returnType` getter (which had complex logic that was
difficult to follow) in terms of two simpler getters:
`declaredReturnType` and `inferredReturnType`.  Also implemented logic
to ensure that inferred instance types are ignored when inferring
types of initializer expressions in the same library cycle (since the
linker is supposed to behave as if initializer expressions are
inferred first).

R=brianwilkerson@google.com, scheglov@google.com

Review URL: https://codereview.chromium.org/1902103002 .