Delay AnalysisContext.exists(source) in DirectiveResolver.
We don't want to report a problem when we have already reported that
the source does not exist earlier. But in the most cases it does exist,
and we are able to find the corresponding import (otherwise we have a
but analyzer). So, we can delay exists() check until right before we
are going to report this bug.

R=brianwilkerson@google.com
BUG=

Review URL: https://codereview.chromium.org/1902783004 .