Modify dump_inferred_types.dart to handle missing unlinked units.
In rare cases a summary package bundle may contain a LinkedLibrary
object but not its corresponding UnlinkedUnits.  (This could happen, for
instance, if both a package and one of its dependencies contain
libraries that reference the same file as a "part"--in this case, the
UnlinkedUnit for the part might only be included in one of the package
bundles, because it is redundant in the other).

This CL modifies dump_inferred_types.dart so that it doesn't crash in
this rare corner case.

R=scheglov@google.com

Review URL: https://codereview.chromium.org/1930663003 .