Remove "@informative" annotation on initializers and local functions.
Given code like this:

var x = (...) => ...;

The inferred type of x is the type of the closure, which is considered
by analyzer to be a local function inside the inside the initializer
expression (which is itself treated as a synthetic function for element
model purposes).  So we can't consider initializers and local functions
to be "informative"; they can affect semantic via inferred types.

R=scheglov@google.com

Review URL: https://codereview.chromium.org/1932873002 .