Restored the MirrorsUsed which was removed in CL 1933283002.
The `git cl land` on https://codereview.chromium.org/1915123008 seemed
to give rise to many timeout failures in the dart2js-jsshell columns of
the build bot; that CL was then reverted. However, it turns out to be
the removal of `@MirrorsUsed(metaTargets: 'ReflectiveTest')` in
https://codereview.chromium.org/1933283002 that causes the timeouts
(e.g., it changes an 11sec test run into a 64s test run, which violates
the timeout constraint of 60s on the bot).

BUG=

Review URL: https://codereview.chromium.org/1936963002 .