Fix for serializing annotations.
Instead of using artificial TypeName nodes we just pass all the required
information. This helps to avoid the situation when we mutate the AST
because we re-parent nodes in annotation.

It would be nice if annotations used the same AST structure as
instance creation :-(

R=paulberry@google.com
BUG=

Review URL: https://codereview.chromium.org/1943013002 .