Adds `DartOptions` to generic method syntax tests (updated CL).
This CL was landed and reverted. The bot failures caused by the first
landing of this CL turned out to be caused by checked mode execution.
This is an adjustment of that CL which marks the relevant tests as
failing. The underlying issue is that the compiler phases after
resolution do not expect to receive method type variables, only class
type variables; that will be addressed in a separate CL.

R=floitsch@google.com, johnniwinther@google.com

Review URL: https://codereview.chromium.org/1963703003 .