always emit lazy field if overriding static
We used to emit an eager field if a static was overriding a static field
in a superclass because this is valid ES6. However, even if the static
field is not synthetic in the superclass, it might still be emitted as a
lazy static. The safest way is just to emit a lazy field for static
overrides always.

fixes #522

R=jmesserly@google.com

Review URL: https://codereview.chromium.org/1966763004 .