Changes dart2js to warn at reified usage of method type variables.
This CL changes the response to `x is T`, `x is! T`, `x as T`, and `T`
used as a type expression (say, `foo(T)`) where `T` is a method type
variable: It will now emit a warning in these cases.

Test examples associated with `TYPE_VARIABLE_FROM_METHOD_NOT_REIFIED`
have been activated and corrected, such that this behavior is tested.

As a result 'language/generic_methods_type_expression_test.dart' will
now cause warnings when compiled with `dart2js`.

Note that this does not break any tests, because `dart2js` does not make
the compilation fail when there is a warning (also, the analyzer and the
vm are unaffected because this is about a change to `dart2js`, not a
change to the test file).

R=johnniwinther@google.com

Review URL: https://codereview.chromium.org/2001093006 .