Make AST-based type inference handle <bottom> consistently.
The non-AST-based type inference engine has the following quirk.  With
the source code:

var x = null;

The type of `x` is inferred as `dynamic`, whereas the type of the
synthetic initializer function associated with `x` is inferred as `()
-> <bottom>`.

This CL makes the AST-based type inference engine act in the same way
for consistency.

In the future we may want to change both engines to consistently infer
`dynamic`.

R=scheglov@google.com

Review URL: https://codereview.chromium.org/2014383003 .